Added profile component, split home page into multiple components, user can now hit enter to submit input fields
The home component was doing a lot of logic and started to feel cluttered, I made the youtube player and song selector into their own components. I also added a directive that allows the user to hit enter when using input fields. A profile page was added, but currently lacks features